Move to engine_common
