+++ mobile/library/common/engine.cc
@@ -7,55 +7,6 @@
 #include "library/common/data/utility.h"
 #include "library/common/stats/utility.h"
 
-#if !defined(ENVOY_ENABLE_FULL_PROTOS)
-
-#include "bazel/cc_proto_descriptor_library/file_descriptor_info.h"
-#include "source/common/protobuf/protobuf.h"
-
-#include "library/common/extensions/cert_validator/platform_bridge/platform_bridge_descriptor.pb.h"
-#include "library/common/extensions/filters/http/local_error/filter_descriptor.pb.h"
-#include "library/common/extensions/filters/http/network_configuration/filter_descriptor.pb.h"
-#include "library/common/extensions/filters/http/platform_bridge/filter_descriptor.pb.h"
-#include "library/common/extensions/filters/http/socket_tag/filter_descriptor.pb.h"
-#include "library/common/extensions/key_value/platform/platform_descriptor.pb.h"
-#include "library/common/extensions/retry/options/network_configuration/predicate_descriptor.pb.h"
-
-namespace Envoy {
-
-bool initialize() {
-  std::vector<FileDescriptorInfo> file_descriptors = {
-      protobuf::reflection::
-          library_common_extensions_cert_validator_platform_bridge_platform_bridge::
-              kFileDescriptorInfo,
-      protobuf::reflection::library_common_extensions_filters_http_local_error_filter::
-          kFileDescriptorInfo,
-      protobuf::reflection::library_common_extensions_filters_http_network_configuration_filter::
-          kFileDescriptorInfo,
-      protobuf::reflection::library_common_extensions_filters_http_platform_bridge_filter::
-          kFileDescriptorInfo,
-      protobuf::reflection::library_common_extensions_filters_http_socket_tag_filter::
-          kFileDescriptorInfo,
-      protobuf::reflection::library_common_extensions_key_value_platform_platform::
-          kFileDescriptorInfo,
-      protobuf::reflection::
-          library_common_extensions_retry_options_network_configuration_predicate::
-              kFileDescriptorInfo,
-  };
-  for (const FileDescriptorInfo& descriptor : file_descriptors) {
-    loadFileDescriptors(descriptor);
-  }
-  return true;
-}
-
-void registerMobileProtoDescriptors() {
-  static bool initialized = initialize();
-  (void)initialized;
-}
-
-} // namespace Envoy
-
-#endif
-
 namespace Envoy {
 
 Engine::Engine(envoy_engine_callbacks callbacks, envoy_logger logger,
@@ -68,10 +19,6 @@ Engine::Engine(envoy_engine_callbacks callbacks, envoy_logger logger,
   // registry may lead to crashes at Engine shutdown. To be figured out as part of
   // https://github.com/envoyproxy/envoy-mobile/issues/332
   Envoy::Api::External::registerApi(std::string(envoy_event_tracker_api_name), &event_tracker_);
-
-#if !defined(ENVOY_ENABLE_FULL_PROTOS)
-  registerMobileProtoDescriptors();
-#endif
 }
 
 envoy_status_t Engine::run(const std::string config, const std::string log_level) {
